requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==4.0.2
django-cors-headers==3.11.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
mysqlclient==2.1.0
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
